Refactor JSON document handling and add extensions
`Corvus.Text.Json.csproj` now includes `JsonElementExtensions.cs`, which provides extension methods for `IJsonElement` - specifically for `CreateBuilder()` functionality. Tthe visibility of the `CreateBuilder` method in `JsonWorkspace` is restricted to internal, and exposed through the extension for `IJsonElement`

In addition, the `Dispose` method in `JsonDocument` is modified to be protected, and `ParsedJsonDocument` now implements `IDisposable`.